requirements.txt
PyPDF2
pdfplumber
PyMuPDF
pdfminer.six
